# server/openapi.yaml
openapi: 3.0.3
info:
  title: Cook Africa API
  description: |
    API backend de l'application de gestion interne du restaurant Cook Africa,
    plus le point d'entrée public utilisé par la page client `/commander`.

    ### Authentification
    La quasi-totalité des routes exigent un JWT Bearer obtenu via `POST /api/auth/login`,
    envoyé dans l'en-tête `Authorization: Bearer <token>`. Le token expire après 12h.

    Les routes sous `/api/public/*` sont volontairement **non authentifiées** : elles sont
    appelées depuis le navigateur d'un client final (page de commande en ligne), pas depuis
    un poste staff. Elles n'exposent que le menu disponible et la création/consultation de
    commandes ; toute autre donnée reste protégée.

    ### Rôles
    `admin`, `caissiere`, `serveur`, `cuisiniere`, `barman`, `caissier-en-ligne` — la plupart des routes
    restreignent l'accès à un sous-ensemble de rôles via un contrôle serveur (403 sinon).

    ### Restriction Wi-Fi
    Pour tout rôle autre que `admin`, les requêtes authentifiées sont en plus filtrées par
    IP si la restriction Wi-Fi est activée (`/api/wifi-config`) — 403 `wifi_restricted`
    si le poste n'est pas sur le réseau autorisé.

    ### Erreurs
    Toutes les erreurs suivent le même format `{ "error": "message" }`, avec le code HTTP
    approprié (400 validation, 401 non authentifié, 403 accès refusé, 404 introuvable,
    409 conflit, 429 limite de requêtes, 500 erreur serveur).
  version: 1.0.0
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/factures:
    get:
      operationId: getApiFactures
      tags: [Factures]
      summary: Liste des factures (staff)
      parameters:
        - name: debut
          in: query
          schema: { type: string, format: date }
        - name: fin
          in: query
          schema: { type: string, format: date }
        - name: statut
          in: query
          schema: { type: string, enum: [payee, partielle] }
      responses:
        '200':
          description: Liste des factures (300 plus récentes, mises en cache 60s)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Facture' }
    post:
      operationId: postApiFactures
      tags: [Factures]
      summary: Générer une facture depuis une commande (admin/caissière)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FactureCreateRequest' }
      responses:
        '201':
          description: Facture générée (statut initial "partielle", TVA déjà incluse dans les prix du menu)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Facture' }
        '400':
          description: Commande vide ou pas encore envoyée à la facturation
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Une facture existe déjà pour cette commande
  /api/factures/{id}:
    get:
      operationId: getApiFacturesById
      tags: [Factures]
      summary: Détail d'une facture (staff)
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Facture
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Facture' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/factures/repair-numeros:
    post:
      operationId: postApiFacturesRepairNumeros
      tags: [Factures]
      summary: Répare les numéros de facture invalides en base (admin, maintenance)
      responses:
        '200':
          description: Factures réparées
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        ancien: { type: string }
                        nouveau: { type: string }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/factures/{id}/pay:
    put:
      operationId: putApiFacturesByIdPay
      tags: [Factures]
      summary: Encaisser le paiement d'une facture (admin/caissière)
      description: |
        La caissière peut ajuster librement les articles/prix avant paiement.
        Décrémente le stock des plats du jour et marque la commande liée `servie`.
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FacturePayRequest' }
      responses:
        '200':
          description: Paiement enregistré
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Facture' }
        '400':
          description: Facture déjà payée
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/factures/{id}/edit-items:
    post:
      operationId: postApiFacturesByIdEditItems
      tags: [Factures]
      summary: Modifier librement les articles d'une facture non payée (admin/caissière)
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FactureEditItemsRequest' }
      responses:
        '200':
          description: Facture modifiée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Facture' }
        '400':
          description: Facture déjà payée ou liste d'articles vide
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Stocks ────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton obtenu via `POST /api/auth/login`, valable 12h.
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      description: Identifiant du document Firestore
      schema: { type: string }
  responses:
...
    Forbidden:
      description: Accès refusé (rôle insuffisant, ou hors réseau Wi-Fi autorisé)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    CommandeItem:
      type: object
      properties:
        menuItemId: { type: string }
        nom: { type: string }
        prix: { type: number }
        quantite: { type: integer }
        sousTotal: { type: number }
        categorie: { type: string }
    CommandeItemInput:
      type: object
      required: [nom, prix, quantite]
      properties:
        menuItemId: { type: string }
        nom: { type: string }
        prix: { type: number }
        quantite: { type: integer, minimum: 1 }
        categorie: { type: string }
...
    Facture:
      type: object
      properties:
        id: { type: string }
        numero: { type: string, example: FACT-0001 }
        type: { type: string, enum: [facture] }
        commandeId: { type: string }
        commandeNumero: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/CommandeItem' }
        tableNumero: { type: string }
        note: { type: string }
        total: { type: number }
        reste: { type: number }
        modePaiement: { type: string, example: especes }
        statut: { type: string, enum: [partielle, payee] }
        serveurNom: { type: string }
        caissiereName: { type: string }
        date: { type: string, format: date }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lastEditedBy: { type: string }
        lastEditedByNom: { type: string }
        lastEditedAt: { type: string, format: date-time }
    FactureCreateRequest:
      type: object
      required: [commandeId]
      properties:
        commandeId: { type: string }
        modePaiement: { type: string, default: especes }
    FacturePayRequest:
      type: object
      properties:
        modePaiement: { type: string }
        items:
          type: array
          description: Optionnel — remplace les articles/prix de la facture avant paiement.
          items: { $ref: '#/components/schemas/CommandeItemInput' }
    FactureEditItemsRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/CommandeItemInput' }
...